Move `TypeDeclStubCreator` from `fuzzyc` passes to `typenodes`
